requirements.txt
aiocontextvars==0.2.2
alembic==1.6.5
argon2-cffi==20.1.0
asn1crypto==0.24.0
async-generator==1.10
attrs==21.2.0
Automat==0.6.0
backcall==0.2.0
bleach==3.3.0
bottle==0.12.19
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cliff==3.8.0
cmaes==0.8.2
cmd2==2.1.2
colorama==0.4.4
colorlog==5.0.1
conda==4.3.16
configobj==5.0.6
configparser==5.0.2
constantly==15.1.0
contextvars==2.4
cryptography==2.1.4
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
docker-pycreds==0.4.0
docutils==0.18.1
dropblock==0.3.0
entrypoints==0.3
gitdb==4.0.7
GitPython==3.1.18
googledrivedownloader==0.4
greenlet==1.1.0
h5py==2.10.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.10
imageio==2.9.0
immutables==0.16
importlib-metadata==4.6.1
incremental==16.10.1
ipykernel==5.5.0
ipython==7.16.1
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
llvmlite
loguru==0.5.3
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.5.1
notebook==6.2.0
numba==0.49.1
numpy==1.19.5
opencv-python==4.2.0.32
optuna==2.8.0
optuna-dashboard==0.4.2
packaging==21.0
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.1
pathtools==0.1.2
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.5.0
plyfile==0.7.1
prettytable==2.1.0
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.16
protobuf==3.17.3
psutil==5.8.0
pthflops==0.4.0
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
Pygments==2.8.1
pykalman==0.9.5
PyOpenGL==3.1.0
pyOpenSSL==17.5.0
pyparsing==3.0.7
pyperclip==1.8.2
pyrsistent==0.17.3
pyserial==3.4
python-dateutil==2.8.2
python-debian==0.1.32
python-editor==1.0.4
pytz==2021.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.4.1
pyzmq==22.0.3
rdflib==4.2.2
requests==2.25.1
requests-unixsocket==0.1.5
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.10.0
SecretStorage==2.3.1
Send2Trash==1.5.0
sentry-sdk==1.3.0
service-identity==16.0.0
shortuuid==1.0.1
six==1.16.0
smmap==4.0.0
SQLAlchemy==1.4.20
stevedore==3.3.0
subprocess32==3.5.4
tabulate==0.8.6
tenacity==8.0.1
termcolor==1.1.0
terminado==0.9.2
terminaltables==3.1.0
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.9.3
timm==0.5.4
torch==1.10.1
torch-geometric==1.4.1
torchvision==0.11.2
tornado==6.1
tqdm==4.61.2
traitlets==4.3.3
Twisted==17.9.0
typing-extensions==4.1.1
urllib3==1.26.6
wandb==0.10.33
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.5.0